+++ src/types/OpenScadSummary.ts
@@ -1,3 +1,24 @@
+import { ParameterDefinition } from "./ParameterDefinition.js";
+
+export interface OpenScadOutput {
+  output: string;
+  modelFile: string;
+  file: string;
+}
+
+export interface ISummary {
+  getArg(): string;
+  getSummary(): ModelSummary;
+}
+
+export interface OpenScadOutputWithSummary extends OpenScadOutput {
+  summary: ModelSummary;
+}
+
+export interface OpenScadOutputWithParameterDefinition extends OpenScadOutput {
+  parameterDefinition: ParameterDefinition;
+}
+
 type Cache = {
   bytes: number;
   entries: number;